pyproject.toml
[tool.poetry]
name = "pinboard-archiver"
version = "0.1.0"
description = "Consumer which archives links send to Pinboard"
authors = ["Simon Lundström <github-commits@soy.se>"]

[tool.poetry.dependencies]
python = "^3.7"
pika = "^1.2.0"
click = "^7.1.2"
pycapnp = "^1.0.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
mypy = "^0.790"
pylint = "^2.6.0"
black = "^20.8b1"

[tool.poetry.scripts]
pinboard-archiver = 'pinboard_archiver:main'

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
